requirements.txt
requests==2.28.1
ansible==5.10.0
mitogen==0.3.3
google-auth==2.17.3
pytest==8.1.1
pytest-testinfra==10.1.0
molecule_gce